+++ bspsrc-decompiler/src/main/java/info/ata4/bspsrc/decompiler/modules/texture/TextureAxis.java
@@ -10,7 +10,7 @@
 
 package info.ata4.bspsrc.decompiler.modules.texture;
 
-import info.ata4.bspsrc.lib.vector.Vector3f;
+import info.ata4.bspsrc.lib.vector.Vector3d;
 
 /**
  * A data structure for UV texture coordinates.
@@ -19,27 +19,27 @@ import info.ata4.bspsrc.lib.vector.Vector3f;
  */
 public class TextureAxis {
 
-    public final Vector3f axis;
-    public final float tw;
+    public final Vector3d axis;
+    public final double tw;
     public final int shift;
 
-    public TextureAxis(Vector3f axis, int shift, float tw) {
+    public TextureAxis(Vector3d axis, int shift, double tw) {
         this.axis = axis;
         this.tw = tw;
         this.shift = shift;
     }
 
-    public TextureAxis(float x, float y, float z, int shift, float tw) {
-        this.axis = new Vector3f(x, y, z);
+    public TextureAxis(double x, double y, double z, int shift, double tw) {
+        this.axis = new Vector3d(x, y, z);
         this.tw = tw;
         this.shift = shift;
     }
 
-    public TextureAxis(Vector3f axis) {
+    public TextureAxis(Vector3d axis) {
         this(axis, 0, 0.25f);
     }
 
-    public TextureAxis(float x, float y, float z) {
+    public TextureAxis(double x, double y, double z) {
         this(x, y, z, 0, 0.25f);
     }
